requirements.txt
numpy
pyserial
minimalmodbus==0.6
